backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
bcrypt
python-dotenv